Seperated face_node into its own file
